% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCall.R
\name{api_call}
\alias{api_call}
\title{place a call to the acalog api}
\usage{
api_call(method)
}
\arguments{
\item{method}{method used for specific requests}
}
\value{
the api payload
}
\description{
place a call to the acalog api
}
