% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendars.R
\name{getCalendars}
\alias{getCalendars}
\title{get a tidy tibble listing the available Calendars on acalog}
\usage{
getCalendars()
}
\value{
tidy tibble of calendars and calendar ids
}
\description{
get a tidy tibble listing the available Calendars on acalog
}
