% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrefixes.R
\name{getPrefixes}
\alias{getPrefixes}
\title{getPrefixes allows you to retrieve a list of course prefixes within a single calendar}
\usage{
getPrefixes(calendar_id)
}
\arguments{
\item{calendar_id}{integer.  Calendar to select}
}
\value{
tibble of calendar_id (dbl) and course_prefix (chr)
}
\description{
getPrefixes allows you to retrieve a list of course prefixes within a single calendar
}
